+++ CHANGELOG.md
@@ -1,5 +1,10 @@
+# v5.2.2
+- Rebased from XMRig v6.8.1 [(changelog)](doc/CHANGELOG.md)
+- Full RandomX integration for Panthera
+- Better algo detection (it only use the "True" CPU Cores, since HT ones only add electricity costs if used) 
+
 # v5.2.1
-- Rebased from XMRig v6.7.0 [(changelog)](doc/CHANGELOG.md)
+- Rebased from XMRig v6.7.0
 - Added Compatibiliy with Apple M1 CPU's
 - Several optimisations for the latest x86 CPU's (mainly Ryzen's)